+++ MyongJunSickFE/.eslintrc.js
@@ -4,10 +4,7 @@ module.exports = {
         "es2021": true,
         "node": true
     },
-    "extends": [
-        "eslint:recommended",
-        "plugin:react/recommended"
-    ],
+    "extends": "plugin:react/recommended",
     "overrides": [
     ],
     "parserOptions": {
