+++ MyongJunSickFE/components/RecoilApp.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+import { RecoilRoot } from "recoil";
+import App from "./App";
+
+export default function RecoilApp() {
+  return (
+    <RecoilRoot>
+      <App />
+    </RecoilRoot>
+  );
+}
